Fix default values assignment for IAM resources
Using Short circuiting for assignment of default value assignment.
