+++ cli/tests/pcluster/templates/test_cdk_builder_utils/test_iam_resource_prefix_build_in_cdk/resourcePrefix.only_path_prefix.yaml
